Add top offset support
We need to support an initial scroll offset at the top of the view.
This is useful for iOS top header translucency where the parallax view
will occupy space behind the header but needs to start with the
content visible. Then as the user scrolls, the content can be seen
sliding, behind the header.